requirements.txt
--find-links=https://girder.github.io/large_image_wheels GDAL
# cartopy
folium==0.13.0
geemap[extra]
geopandas
jupyter-server-proxy
keplergl
leafmap
localtileserver
nbserverproxy
owslib
palettable
plotly
streamlit
streamlit-bokeh-events
streamlit-folium
streamlit-keplergl
tropycal
leafmap[lidar]
laspy
segment-geospatial
torch
localtileserver
osmnx
rasterstats
prophet
streamlit_image_select
prettymapp
# git+https://github.com/giswqs/leafmap
# git+https://github.com/giswqs/geemap

